+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
  * 0.0.8 _Jul.15.2023_
    * use scandir filter function param
+   * use bold for directory names
  * 0.0.7 _Jul.09.2023_
    * use file-detail alist at sorting
    * use filtered file list, without '.' and '..'